+++ src/message.rs
@@ -4,9 +4,8 @@ use std::string::String;
 use std::collections::BTreeMap;
 use std::convert::Into;
 use std::ops;
-
-#[cfg(feature = "rustc-serialize")]
-use rustc_serialize::{Encodable,Encoder};
+use std::cmp::Ordering;
+use std::str::FromStr;
 
 #[cfg(feature="serde-serialize")]
 use serde::{Serializer, Serialize};
@@ -17,38 +16,48 @@ pub type time_t = i64;
 pub type pid_t = i32;
 #[allow(non_camel_case_types)]
 pub type msgid_t = String;
-pub type MessageType = String;
 
 use severity;
 use facility;
+use parser;
 
 
 #[derive(Clone,Debug,PartialEq,Eq)]
 /// `ProcID`s are usually numeric PIDs; however, on some systems, they may be something else
-pub enum ProcIdType {
+pub enum ProcId {
     PID(pid_t),
-    Name(String)
+    Name(String),
+}
+
+
+impl PartialOrd for ProcId {
+    fn partial_cmp(&self, other: &ProcId) -> Option<Ordering> {
+        match (self, other) {
+            (&ProcId::PID(ref s_p), &ProcId::PID(ref o_p)) => Some(s_p.cmp(o_p)),
+            (&ProcId::Name(ref s_n), &ProcId::Name(ref o_n)) => Some(s_n.cmp(o_n)),
+            _ => None,
+        }
+    }
 }
 
 
 #[cfg(feature = "rustc-serialize")]
-impl Encodable for ProcIdType {
-    fn encode<S: Encoder>(&self, s: &mut S) -> Result<(), S::Error>
-    {
+impl Encodable for ProcId {
+    fn encode<S: Encoder>(&self, s: &mut S) -> Result<(), S::Error> {
         match *self {
-            ProcIdType::PID(ref p) => s.emit_i32(*p),
-            ProcIdType::Name(ref n) => s.emit_str(n)
+            ProcId::PID(ref p) => s.emit_i32(*p),
+            ProcId::Name(ref n) => s.emit_str(n),
         }
     }
 }
 
 
 #[cfg(feature = "serde-serialize")]
-impl Serialize for ProcIdType {
+impl Serialize for ProcId {
     fn serialize<S: Serializer>(&self, ser: S) -> Result<S::Ok, S::Error> {
         match *self {
-            ProcIdType::PID(ref p) => ser.serialize_i32(*p),
-            ProcIdType::Name(ref n) => ser.serialize_str(n),
+            ProcId::PID(ref p) => ser.serialize_i32(*p),
+            ProcId::Name(ref n) => ser.serialize_str(n),
         }
     }
 }
@@ -85,8 +94,7 @@ impl ops::Deref for StructuredData {
 
 #[cfg(feature = "rustc-serialize")]
 impl Encodable for StructuredData {
-    fn encode<S: Encoder>(&self, s: &mut S) -> Result<(), S::Error>
-    {
+    fn encode<S: Encoder>(&self, s: &mut S) -> Result<(), S::Error> {
         self.elements.encode(s)
     }
 }
@@ -99,24 +107,29 @@ impl Serialize for StructuredData {
 }
 
 impl StructuredData {
-    pub fn new_empty() -> Self
-    {
-        StructuredData {
-            elements: BTreeMap::new()
-        }
+    pub fn new_empty() -> Self {
+        StructuredData { elements: BTreeMap::new() }
     }
 
     /// Insert a new (sd_id, sd_param_id) -> sd_value mapping into the StructuredData
-    pub fn insert_tuple<SI, SPI, SPV> (&mut self, sd_id: SI, sd_param_id: SPI, sd_param_value: SPV) -> ()
-        where SI: Into<SDIDType>, SPI: Into<SDParamIDType>, SPV: Into<SDParamValueType>
+    pub fn insert_tuple<SI, SPI, SPV>(&mut self,
+                                      sd_id: SI,
+                                      sd_param_id: SPI,
+                                      sd_param_value: SPV)
+                                      -> ()
+        where SI: Into<SDIDType>,
+              SPI: Into<SDParamIDType>,
+              SPV: Into<SDParamValueType>
     {
         let sub_map = self.elements.entry(sd_id.into()).or_insert_with(BTreeMap::new);
         sub_map.insert(sd_param_id.into(), sd_param_value.into());
     }
 
     /// Lookup by SDID, SDParamID pair
-    pub fn find_tuple<'b>(&'b self, sd_id: &str, sd_param_id: &str) -> Option<&'b SDParamValueType>
-    {
+    pub fn find_tuple<'b>(&'b self,
+                          sd_id: &str,
+                          sd_param_id: &str)
+                          -> Option<&'b SDParamValueType> {
         // TODO: use traits to make these based on the public types isntead of &str
         if let Some(sub_map) = self.elements.get(sd_id) {
             if let Some(value) = sub_map.get(sd_param_id) {
@@ -130,28 +143,25 @@ impl StructuredData {
     }
 
     /// Find all param/value mappings for a given SDID
-    pub fn find_sdid<'b>(&'b self, sd_id: &str) -> Option<&'b BTreeMap<SDParamIDType, SDParamValueType>>
-    {
+    pub fn find_sdid<'b>(&'b self, sd_id: &str) -> Option<&'b StructuredDataElement> {
         self.elements.get(sd_id)
     }
 
     /// The number of distinct SD_IDs
-    pub fn len(&self) -> usize
-    {
+    pub fn len(&self) -> usize {
         self.elements.len()
     }
 
     /// Whether or not this is empty
-    pub fn is_empty(&self) -> bool
-    {
+    pub fn is_empty(&self) -> bool {
         self.elements.is_empty()
     }
 }
 
 
-#[cfg_attr(feature = "rustc-serialize", derive(RustcEncodable))]
 #[cfg_attr(feature = "serde-serialize", derive(Serialize))]
-#[derive(Clone,Debug)]
+#[derive(Clone,Debug,PartialEq,Eq)]
+/// A RFC5424-protocol syslog message
 pub struct SyslogMessage {
     pub severity: severity::SyslogSeverity,
     pub facility: facility::SyslogFacility,
@@ -159,25 +169,34 @@ pub struct SyslogMessage {
     pub timestamp: Option<time_t>,
     pub hostname: Option<String>,
     pub appname: Option<String>,
-    pub procid: Option<ProcIdType>,
+    pub procid: Option<ProcId>,
     pub msgid: Option<msgid_t>,
     pub sd: StructuredData,
     pub msg: String,
 }
 
 
+impl FromStr for SyslogMessage {
+    type Err = parser::ParseErr;
+
+    /// Parse a string into a `SyslogMessage`
+    ///
+    /// Just calls `parser::parse_message`
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        parser::parse_message(s)
+    }
+}
+
+
 #[cfg(test)]
 mod tests {
-    #[cfg(feature = "rustc-serialize")]
-    use rustc_serialize;
     #[cfg(feature = "serde-serialize")]
     use serde_json;
     use super::StructuredData;
-    #[cfg(any(feature="serde-serialize", feature="rustc-serialize"))]
     use super::SyslogMessage;
-    #[cfg(any(feature="serde-serialize", feature="rustc-serialize"))]
+    #[cfg(feature="serde-serialize")]
     use severity::SyslogSeverity::*;
-    #[cfg(any(feature="serde-serialize", feature="rustc-serialize"))]
+    #[cfg(feature="serde-serialize")]
     use facility::SyslogFacility::*;
 
     #[test]
@@ -189,17 +208,6 @@ mod tests {
         assert!(s.find_tuple("foo", "baz").is_none());
     }
 
-    #[cfg(feature = "rustc-serialize")]
-    #[test]
-    fn test_structured_data_serialization_rustc_serialize() {
-        let mut s = StructuredData::new_empty();
-        s.insert_tuple("foo", "bar", "baz");
-        s.insert_tuple("foo", "baz", "bar");
-        s.insert_tuple("faa", "bar", "baz");
-        let encoded = rustc_serialize::json::encode(&s).expect("Should encode to JSON");
-        assert_eq!(encoded, r#"{"faa":{"bar":"baz"},"foo":{"bar":"baz","baz":"bar"}}"#);
-    }
-
     #[cfg(feature = "serde-serialize")]
     #[test]
     fn test_structured_data_serialization_serde() {
@@ -208,30 +216,8 @@ mod tests {
         s.insert_tuple("foo", "baz", "bar");
         s.insert_tuple("faa", "bar", "baz");
         let encoded = serde_json::to_string(&s).expect("Should encode to JSON");
-        assert_eq!(encoded, r#"{"faa":{"bar":"baz"},"foo":{"bar":"baz","baz":"bar"}}"#);
-    }
-
-    #[cfg(feature = "rustc-serialize")]
-    #[test]
-    fn test_serialization_rustc_serialize() {
-        let m = SyslogMessage {
-            severity: SEV_INFO,
-            facility: LOG_KERN,
-            version: 1,
-            timestamp: None,
-            hostname: None,
-            appname: None,
-            procid: None,
-            msgid: None,
-            sd: StructuredData::new_empty(),
-            msg: String::from("")
-        };
-
-        let encoded = rustc_serialize::json::encode(&m).expect("Should encode to JSON");
-        println!("{:?}", encoded);
-        // XXX: we don't have a guaranteed order, I don't think, so this might break with minor
-        // version changes. *shrug*
-        assert_eq!(encoded, "{\"severity\":\"info\",\"facility\":\"kern\",\"version\":1,\"timestamp\":null,\"hostname\":null,\"appname\":null,\"procid\":null,\"msgid\":null,\"sd\":{},\"msg\":\"\"}");
+        assert_eq!(encoded,
+                   r#"{"faa":{"bar":"baz"},"foo":{"bar":"baz","baz":"bar"}}"#);
     }
 
     #[cfg(feature = "serde-serialize")]
@@ -247,14 +233,14 @@ mod tests {
             procid: None,
             msgid: None,
             sd: StructuredData::new_empty(),
-            msg: String::from("")
+            msg: String::from(""),
         };
 
         let encoded = serde_json::to_string(&m).expect("Should encode to JSON");
-        println!("{:?}", encoded);
         // XXX: we don't have a guaranteed order, I don't think, so this might break with minor
         // version changes. *shrug*
-        assert_eq!(encoded, "{\"severity\":\"info\",\"facility\":\"kern\",\"version\":1,\"timestamp\":null,\"hostname\":null,\"appname\":null,\"procid\":null,\"msgid\":null,\"sd\":{},\"msg\":\"\"}");
+        assert_eq!(encoded,
+                   "{\"severity\":\"info\",\"facility\":\"kern\",\"version\":1,\"timestamp\":null,\"hostname\":null,\"appname\":null,\"procid\":null,\"msgid\":null,\"sd\":{},\"msg\":\"\"}");
     }
 
     #[test]
@@ -266,5 +252,11 @@ mod tests {
         assert_eq!("baz", s.get("foo").and_then(|foo| foo.get("bar")).unwrap());
         assert_eq!("bar", s.get("foo").and_then(|foo| foo.get("baz")).unwrap());
         assert_eq!("baz", s.get("faa").and_then(|foo| foo.get("bar")).unwrap());
+    }
+
+    #[test]
+    fn test_fromstr() {
+        let msg = "<1>1 1985-04-12T23:20:50.52Z host - - - -".parse::<SyslogMessage>().expect("Should parse empty message");
+        assert_eq!(msg.timestamp, Some(482196050));
     }
 }